+++ src/main/java/io/github/mapepire_ibmi/types/exceptions/UnknownServerException.java
@@ -1,4 +1,4 @@
-package io.github.mapepire_ibmi;
+package io.github.mapepire_ibmi.types.exceptions;
 
 /**
  * Represents an unknown exception from the Mapepire server.
